✨ 소유권 검사 annotation 수정

+++ src/main/java/com/example/gemm_server/domain/repository/GemRepository.java
@@ -4,6 +4,6 @@ import com.example.gemm_server.domain.entity.Gem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface GemRepository extends JpaRepository<Gem, Integer> {
+public interface GemRepository extends JpaRepository<Gem, Long> {
 
 }
